+++ lib/td_dd/data_structures/data_structures.ex
@@ -1080,11 +1080,6 @@ defmodule TdDd.DataStructures do
     end)
   end
 
-  import Ecto.Query, warn: false
-  alias TdDd.Repo
-
-  alias TdDd.DataStructures.StructureNote
-
   @doc """
   Returns the list of structure_notes.