+++ lib/td_dd/data_structures/profiling_loader.ex
@@ -5,7 +5,7 @@ defmodule TdDd.ProfilingLoader do
 
   alias TdDd.DataStructures
   alias TdDd.DataStructures.DataStructure
-  alias TdDd.DataStructures.Profiles
+  alias TdDd.Profiles
   alias TdDd.Repo
 
   require Logger